# agentscope-service/docs/service-api/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentScope Service Invocation API",
    "version": "1.0.0"
  },
  "security": [
    {
      "endpointKey": []
    },
    {
      "platformBearer": []
    }
  ],
  "paths": {
...
    "/api/v1/applications": {
      "get": {
        "summary": "List applications",
        "operationId": "get_api_v1_applications",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "platformBearer": []
          }
        ]
      },
      "post": {
        "summary": "Create application",
        "operationId": "post_api_v1_applications",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "application": {
                      "$ref": "#/components/schemas/Application"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "platformBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationInput"
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{applicationId}": {
      "get": {
        "summary": "Read application",
        "operationId": "get_api_v1_applications_applicationId",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "platformBearer": []
          }
        ]
      },
      "patch": {
        "summary": "Update application",
        "operationId": "patch_api_v1_applications_applicationId",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "platformBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationPatch"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "endpointKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "platformBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "registrationBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Scoped Agent registration credential, not an Endpoint credential"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "snapshot_url": {
            "type": "string"
          }
        }
      },
      "ApplicationMember": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "minItems": 1,
            "items": {
              "enum": [
                "viewer",
                "operator",
                "approver"
              ]
            }
          }
        },
        "required": [
          "userId",
          "roles"
        ]
      },
      "ApplicationInput": {
        "type": "object",
        "properties": {
          "tenant": {
            "type": "string"
          },
          "namespace": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApplicationMember"
            }
          },
          "maxConcurrent": {
            "type": "integer",
            "minimum": 0
          },
          "tokenBudget": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "tenant",
          "namespace",
          "name"
        ]
      },
      "Application": {
        "type": "object",
        "properties": {
          "tenant": {
            "type": "string"
          },
          "namespace": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApplicationMember"
            }
          },
          "maxConcurrent": {
            "type": "integer",
            "minimum": 0
          },
          "tokenBudget": {
            "type": "integer",
            "minimum": 0
          },
          "id": {
            "type": "string"
          },
          "ownerUserId": {
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "disabled"
            ]
          },
          "version": {
            "type": "integer"
          },
          "tokensUsed": {
            "type": "integer",
            "readOnly": true
          }
        }
      },
      "ApplicationPatch": {
        "type": "object",
        "properties": {
          "tenant": {
            "type": "string"
          },
          "namespace": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApplicationMember"
            }
          },
          "maxConcurrent": {
            "type": "integer",
            "minimum": 0
          },
          "tokenBudget": {
            "type": "integer",
            "minimum": 0
          },
          "version": {
            "type": "integer"
          },
          "status": {
            "enum": [
              "active",
              "disabled"
            ]
          }
        },
        "required": [
          "version"
        ]
      },
...
    }
  }
}